Refactor: Dissolve Aggs class
